Implement TransactionCache.getKeys()
